[5.x] Extract whereSite query method to trait (#9991)
